Explicitly specify the right module AID in usage instructions
The module AID of our CAP files has changed, and various tools erroneously use
it as the applet AID. This cannot be fixed because the applet AID can not
be specified in a CAP file.

This should fix issue #10, though not in a horribly nice way.

We are still considering if moving back to the default AID is an option,
but currently we would like to stick to proper AIDs for module naming.